refactor : fallback메서드 분리
