+++ couponmoa-coupon/src/main/java/com/couponmoa/backend/couponmoacoupon/domain/coupon/enums/CouponStatus.java
@@ -23,7 +23,3 @@ public enum CouponStatus {
     }
 }
 
-// 생성 시점에는 이미 날짜 검증 로직이 존재하기 때문에,
-// UPCOMING 상태만 유효함.
-// 이후에는 시간의 흐름에 따라 자동으로 변경 or 업데이트 시 자동으로 변경( 카테고리를 직접 변경할 수는 없음)
-// + 조회해서 응답값으로 반환할때도 변경이 되도록 로직 추가. 뭔가 더해야할까 ?
