+++ couponmoa-coupon/src/main/java/com/couponmoa/backend/couponmoacoupon/domain/subscribe/usercoupon/controller/UserCouponSubscribeController.java
@@ -16,13 +16,13 @@ import java.util.List;
 @RequestMapping("/api/v1/coupons")
 public class UserCouponSubscribeController {
 
-    private final UserCouponSubscribeService userCouponSubServ;
+    private final UserCouponSubscribeService userCouponSubscribeService;
 
     @PostMapping("/{couponId}/subscriptions")
     public ResponseEntity<ApiResponse<Void>> subscribeCoupon(@RequestHeader("X-User-Id") Long userId,
                                                              @Parameter(description = "구독할 쿠폰 id", required = true)
                                                              @PathVariable Long couponId) {
-        userCouponSubServ.subscribeCoupon(userId, couponId);
+        userCouponSubscribeService.subscribeCoupon(userId, couponId);
 
         return ResponseEntity.ok(ApiResponse.success(couponId + "번 쿠폰 구독 완료"));
     }
@@ -31,7 +31,7 @@ public class UserCouponSubscribeController {
     public ResponseEntity<ApiResponse<Void>> unsubscribeCoupon(@RequestHeader("X-User-Id") Long userId,
                                                                @Parameter(description = "구독할 쿠폰 id", required = true)
                                                                @PathVariable Long couponId) {
-        userCouponSubServ.unSubscribeCoupon(userId, couponId);
+        userCouponSubscribeService.unSubscribeCoupon(userId, couponId);
 
         return ResponseEntity.ok(ApiResponse.success(couponId + "번 쿠폰 구독 취소"));
     }
@@ -40,14 +40,14 @@ public class UserCouponSubscribeController {
     public ResponseEntity<ApiResponse<List<FindCouponSubscribeListResponse>>> findCouponSubscribeList(@RequestHeader("X-User-Id") Long userId,
                                                                                                       @RequestParam(defaultValue = "0") int page,
                                                                                                       @RequestParam(defaultValue = "10") int size) {
-        List<FindCouponSubscribeListResponse> subscribeList = userCouponSubServ.findSubscribeList(userId, page, size);
+        List<FindCouponSubscribeListResponse> subscribeList = userCouponSubscribeService.findSubscribeList(userId, page, size);
 
         return ResponseEntity.ok(ApiResponse.success(subscribeList, "쿠폰 구독 목록 조회 성공"));
     }
 
     @PostMapping("/{couponId}/alert")
     public ResponseEntity<ApiResponse<List<String>>> sendAlert(@PathVariable Long couponId) {
-        List<String> emailList = userCouponSubServ.sendAlert(couponId);
+        List<String> emailList = userCouponSubscribeService.sendAlert(couponId);
 
         return ResponseEntity.ok(ApiResponse.success(emailList));
     }
